loomX: interprocedural pointer-write analysis, reduction fixes, GPU speedup
- Allow per-iteration disjoint pointer writes through helpers when the
  index is the loop iterator; add --intraprocedural-baseline flag.
- Fix reduction detector to ignore loop-local temporaries.
- Fix private-variable collection to skip variables declared inside the
  loop body.
- Count nested-loop operations by trip count in compute-intensity
  estimation; use 64-bit counters to avoid overflow.
- Classify outer loops with large nested FLOP work as GPU_OFFLOAD.
- Add validation tests for disjoint pointer writes and loop-local
  reductions; PolyBench gemm now shows 4.4x GPU speedup.

+++ tools/loomX/src/LoopAnalysisTypes.h
@@ -93,11 +93,11 @@ enum class IntensityClass {
 // Result of compute-intensity estimation.
 struct ComputeIntensityResult {
     IntensityClass classification = IntensityClass::UNKNOWN;
-    double flopsPerMemoryOp = 0.0;  // FLOPs per memory access
-    int flopCount = 0;              // Static estimate of FP operations
-    int memoryOpCount = 0;          // Static estimate of memory accesses
-    int integerOpCount = 0;         // Static estimate of integer operations
-    bool hasHeavyMath = false;      // sin/cos/sqrt/exp/log/etc.
+    double flopsPerMemoryOp = 0.0;   // FLOPs per memory access
+    long long flopCount = 0;         // Static estimate of FP operations
+    long long memoryOpCount = 0;     // Static estimate of memory accesses
+    long long integerOpCount = 0;    // Static estimate of integer operations
+    bool hasHeavyMath = false;       // sin/cos/sqrt/exp/log/etc.
     std::string note;
 };
